requirements.txt
allure-pytest==2.10.0
Appium-Python-Client==1.3.0
pytest==7.1.3
selenium==3.141.0
applescript==2021.2.9